Refactor parameter types in Figura262E15 methods and update related files for consistency

+++ src/Application/Class1.cs
@@ -1,8 +1,8 @@
 ﻿public class Figura262E15 {
-public static double Perimetro(double lado, double numeroLados) {
+public static double Perimetro(double lado, int numeroLados) {
 return lado * numeroLados;
 }
-public static double Area(double lado, double apotema, double numeroLados) {
+public static double Area(double lado, double apotema, int numeroLados) {
 double perimetro = Perimetro(lado, numeroLados);
 return perimetro * apotema/2;
 }
